Fix: Add default case to getHostingPriority to prevent undefined return
The switch statement in getHostingPriority was missing a default case,
which could cause the function to return undefined if an unknown provider
type is received from the backend. This would result in NaN during
sorting arithmetic, causing unpredictable sort order.

Added 'default: return 3' to ensure unknown providers are placed at
the end of the sorted list, consistent with the pattern used in the
shared provider-utils.ts utility.

Applied via @cursor push command
