+++ ayunis-core-frontend/src/widgets/model-type-card/index.ts
@@ -1,0 +1,2 @@
+export { default as ModelTypeCard } from './ModelTypeCard';
+export type { ModelActions } from './ModelTypeCard';
