+++ ayunis-core-frontend/src/widgets/subscription-cancellation-section/index.ts
@@ -1,0 +1,2 @@
+export { default as SubscriptionCancellationSection } from './SubscriptionCancellationSection';
+export type { SubscriptionCancellationActions } from './SubscriptionCancellationSection';
